+++ src/hooks/useSignMessage.tsx
@@ -49,7 +49,7 @@ export const useSignMessage = () => {
 
       if (networkStateSnap.dappProvider instanceof ExtensionProvider) {
         const signedMessage = await networkStateSnap.dappProvider.signMessage(
-          // TODO: replace SignableMessage when signing providers are ready
+          // TODO: SignableMessage is deprecated, replace when signing providers are ready
           new SignableMessage({ message: Buffer.from(message) })
         );
 
@@ -57,6 +57,7 @@ export const useSignMessage = () => {
       }
       if (networkStateSnap.dappProvider instanceof WalletConnectV2Provider) {
         const signedMessage = await networkStateSnap.dappProvider.signMessage(
+          // TODO: SignableMessage is deprecated, replace when signing providers are ready
           new SignableMessage({ message: Buffer.from(message) })
         );
 
@@ -64,6 +65,7 @@ export const useSignMessage = () => {
       }
       if (networkStateSnap.dappProvider instanceof HWProvider) {
         const signedMessage = await networkStateSnap.dappProvider.signMessage(
+          // TODO: SignableMessage is deprecated, replace when signing providers are ready
           new SignableMessage({ message: Buffer.from(message) })
         );
 
@@ -81,6 +83,7 @@ export const useSignMessage = () => {
           `${window.location.origin}${options?.callbackUrl || ''}` ||
           window?.location.href;
         await networkStateSnap.dappProvider.signMessage(
+          // TODO: SignableMessage is deprecated, replace when signing providers are ready
           new SignableMessage({ message: Buffer.from(message) }),
           {
             callbackUrl: encodeURIComponent(